requirements.txt
asttokens==2.4.1
cycler==0.11.0
executing==2.0.1
fonttools==4.38.0
gputil==1.4.0
icecream==2.1.3
imageio==2.31.2
imageio-ffmpeg==0.5.1
kiwisolver==1.4.5
markdown==3.3.3
matplotlib==3.5.3
opencv-python==4.9.0.80
packaging==24.0
pandas==1.3.5
psutil==5.9.8
pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
torch-tb-profiler==0.4.3
